+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - Menus Listing after clicking a single Catefory
 ### Fixed
 - createdAt/updatedAt remove default val [#reason](default val is autogen when using upsert/update in sequelize)
+- Register layout
 
 ## 0.0.1 - 2018-09-17
 ### Fixed